frontend: ArticleTable: refactor filter sidebar.
